+++ src/atomist.config.ts
@@ -15,52 +15,49 @@
  */
 
 import { Configuration } from "@atomist/automation-client";
+import { AutomationEventListener } from "@atomist/automation-client/server/AutomationEventListener";
 import * as appRoot from "app-root-path";
 import * as config from "config";
 
-import { LogzioOptions } from "./util/logzio";
-import { secret } from "./util/secrets";
+import {
+    resolveApplication,
+    resolveEnvironment,
+    resolveTeamIds,
+    resolveToken,
+} from "./util/configuration";
 
 // tslint:disable-next-line:no-var-requires
-const pj = require(`${appRoot}/package.json`);
+const pj = require(`${appRoot.path}/package.json`);
 
-const token = secret("github.token", process.env.GITHUB_TOKEN);
 const notLocal = process.env.NODE_ENV === "production" || process.env.NODE_ENV === "staging";
 
-const logzioOptions: LogzioOptions = {
-    applicationId: secret("applicationId"),
-    environmentId: secret("environmentId"),
-    token: secret("logzio.token", process.env.LOGZIO_TOKEN),
-};
+export interface ConfigurationPlus {
+    application?: string;
+    environment?: string;
+    groups?: string[];
+}
 
-export const configuration: Configuration = {
+export const configuration: Configuration & ConfigurationPlus = {
     name: pj.name,
     version: pj.version,
     keywords: pj.keywords,
+    environment: resolveEnvironment(),
+    application: resolveApplication(),
     policy: config.get("policy"),
-    teamIds: config.get("teamIds"),
-    token,
-    http: {
-        enabled: true,
-        auth: {
-            basic: {
-                enabled: config.get("http.auth.basic.enabled"),
-                username: secret("dashboard.user"),
-                password: secret("dashboard.password"),
-            },
-            bearer: {
-                enabled: config.get("http.auth.bearer.enabled"),
-                adminOrg: "atomisthq",
-            },
-        },
-    },
+    teamIds: resolveTeamIds(),
+    groups: config.get("groups"),
+    token: resolveToken(),
     endpoints: {
         api: config.get("endpoints.api"),
         graphql: config.get("endpoints.graphql"),
     },
+    listeners: [],
+    http: {
+        enabled: true,
+    },
     applicationEvents: {
         enabled: true,
-        teamId: "T29E48P34",
+        teamId: process.env.ATOMIST_TEAM,
     },
     cluster: {
         enabled: notLocal,
@@ -72,4 +69,3 @@ export const configuration: Configuration = {
         },
     },
 };
-(configuration as any).groups = config.get("groups");