+++ src/index.d.ts
@@ -39,6 +39,3 @@ declare module "tmp-promise" {
     }
     export function dir(options?: any): Promise<TmpDir>;
 }
-
-declare module "logzio-nodejs";
-declare module "serialize-error";